+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## master (unreleased)
 
 ### Bugs fixed
+
 * [#374](https://github.com/bbatsov/rubocop/issues/374) - Fixed error at post condition loop (`begin-end-while`, `begin-end-until`) in `UnusedLocalVariable` and `ShadowingOuterLocalVariable`
+* [#373](https://github.com/bbatsov/rubocop/issues/373) and [#376](https://github.com/bbatsov/rubocop/issues/376) - allow braces around multi-line blocks if `do`-`end` would change the meaning of the code
 
 ## 0.10.0 (17/07/2013)